2026/03/26 3.30

+++ src/main/java/com/example/CakeShopManagement/entity/CategoryEntity.java
@@ -11,30 +11,27 @@ public class CategoryEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int categoryId;
+    private Long categoryId;
     private String categoryName;
 
     @Lob
     private String description;
 
-//    @ManyToMany(mappedBy = "categories")
-//    private List<ProductEntity> products;
-
 
     public CategoryEntity() {
     }
 
-    public CategoryEntity(int categoryId, String categoryName, String description) {
+    public CategoryEntity(Long categoryId, String categoryName, String description) {
         this.categoryId = categoryId;
         this.categoryName = categoryName;
         this.description = description;
     }
 
-    public int getCategoryId() {
+    public Long getCategoryId() {
         return categoryId;
     }
 
-    public void setCategoryId(int categoryId) {
+    public void setCategoryId(Long categoryId) {
         this.categoryId = categoryId;
     }
 
@@ -54,11 +51,4 @@ public class CategoryEntity {
         this.description = description;
     }
 
-//    public List<ProductEntity> getProducts() {
-//        return products;
-//    }
-//
-//    public void setProducts(List<ProductEntity> products) {
-//        this.products = products;
-//    }
 }
